+++ pyrinnaitouch/heater.py
@@ -2,9 +2,8 @@
 import logging
 from multiprocessing import managers
 from tempfile import tempdir
-from .system import Zone
 
-from .util import get_attribute, y_n_to_bool, symbol_to_schedule_period
+from .util import get_attribute, y_n_to_bool, symbol_to_schedule_period, Zone
 
 _LOGGER = logging.getLogger(__name__)
