+++ pyrinnaitouch/util.py
@@ -35,3 +35,34 @@ def symbol_to_schedule_period(symbol):
     if symbol == "S":
         return SchedulePeriod.SLEEP
     return None
+
+class Zone():
+
+    def __init__(self, name) -> None:
+        self.name = name
+
+    name = ""
+    temperature = 999
+    set_temp = 0 # < 8 means off
+    schedule_period = None
+    advance_period = None
+    advanced = False
+    user_enabled = False # applies only to fan_only
+    auto_mode = False
+
+    def set_mode(self,mode):
+        """Set auto/manual mode."""
+        # A = Auto Mode and M = Manual Mode
+        if mode == "A":
+            self.auto_mode = True
+        elif mode == "M":
+            self.auto_mode = False
+
+    def set_advanced(self,status_str):
+        """Set advanced state."""
+        # A = Advance, N = None, O = Operation (what is that?)
+        if status_str == "A":
+            self.advanced = True
+        else:
+            self.advanced = False
+
